pyproject.toml
[project]
name = "openrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiopath>=0.7.7",
    "chainlit>=2.2.1",
    "docling>=2.24.0",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.39",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.1.2",
    "langchain-milvus>=0.1.8",
    "langchain-openai>=0.3.7",
    "loguru>=0.7.3",
    "marker-pdf>=0.2.17",
    "pydub>=0.25.1",
    "pymupdf4llm>=0.0.17",
    "spire-doc>=13.1.0",
    "openai>=1.64.0",
    "ray[default]>=2.47.1",
    "langchain-qdrant>=0.2.0",
    "python-dotenv>=1.0.1",
    "pip>=25.0.1",
    "chardet>=5.2.0",
    "torch>=2.4.1",
    "robotframework>=7.2.2",
    "robotframework-requests>=0.9.7",
    "infinity-client>=0.0.76",
    "asyncpg>=0.30.0",
    "boto3>=1.38.33",
    "eml-parser>=2.0.0",
    "psutil>=7.0.0",
    "sqlalchemy_utils",
    "psycopg2>=2.9.10",
    "llvmlite>=0.44.0",
    "numba>=0.61.2",
    "umap-learn>=0.5.9.post2",
    "hdbscan>=0.8.40",
    "pytest-env>=1.1.5",
    "markitdown[docx]>=0.1.3",
    "html-to-markdown>=2.4.0",
    "langdetect>=1.0.9",
    "alembic>=1.17.0",
    "langdetect>=1.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
lint = [
    "ruff>=0.14.1",
]
